Now displaying actual timer info in main view’s rows.
Added a row for showing “And N more” when we can’t show all the timers.

Added a global timer database constant, timerDB.

Added API to register for callbacks when the value of a timer changes in the database. The database class and the UI are both imperative shells. We try to keep everything functional inside that shell.