Made TimerData’s updateTimer insert the timer if it’s a new one.
Added support for database reload callback to allow the main interface to reload when a timer is added. We could make this finer grained, but a full reload seems easiest for now and does the trick.

Table reload has to be lazy so that it only happens when we’re active. Sigh.