Implemented the continuation to handle results from commands sent to the iPhone app. Now updating the current timers when new timers are returned.
Added a private backing store for timers, a public getter, and a private setter.

The setter will merge incoming timers from the iPhone into the current set, invoking appropriate callbacks and maintaining order as needed (though it doesn’t do all of that yet). The setter also maintains the timer index and eventually the timerTimers as well.